requirements.txt
python-telegram-bot>=20.6
requests>=2.30.0
python-dotenv>=1.0.0
aiohttp>=3.8.6
cloudscraper>=1.2.71
beautifulsoup4>=4.12.2
matplotlib>=3.7.3
pillow>=9.5.0
openai>=1.10.0 